Extract Nuxt config in utils file
