+++ custom.dic
@@ -1,47 +1,67 @@
 adminsdk
 antfu
+betawerte
+betawerten
 binomiale
 binomialer
 bmiq
 bono
 breitformatiges
+calin
 calina
+cholangiocarcinoma
+cholangiokarzinom
 coderabbit
 composables
 dragomir
 dropfile
+duktaladenokarzinom
 eavatar
 emojione
+ergebnisse
 ffpe
 firestore
 gehostet
+genomweiten
 germany
+goeppert
 gserviceaccount
 hellmodus
 idat
+impressum
 interpretability
 intrahepatic
 intrahepatische
+intrahepatischem
+intrahepatischem
 klassifikationsprozess
 klassifikatordienstes
+kontakt
 methylation
 methylierungsbasierter
 methylierungsdaten
 methylierungsprofile
+methylierungsprofilierung
 nuxi
 nuxt
 paad
 packagename
 pancreato
+pankreatischem
 pankreatobiliärtumore
 registrierungsinformationen
+roessler
+roßner
 samplesheet
+schallenberg
 scorei
+sers
 sonner
 streitbeilegungsverfahren
 teodor
 uncategorized
 uuidv
+verifizierung
 vorhersagedetails
 vorhersageetikett
 vorhersageetiketten
